+++ backend/.env
@@ -6,7 +6,7 @@ DISTANCE_THRESHOLD_MEDIUM=0.35
 DISTANCE_THRESHOLD_LARGE=0.25
 SECRET_KEY=your-secret-key-here-for-development-only
 FRONTEND_URL=http://localhost:5173
-GMAIL_SENDER_EMAIL=some-email@gmail.com						# Replace with your actual Gmail, or use mine sorakh2756@gmail.com
-GMAIL_APP_PASSWORD=gmailAppPassword							# Replace with your actual App Password, or use mine azktsjaohgvppdxo
+GMAIL_SENDER_EMAIL=sorakh2756@gmail.com						# Replace with your actual Gmail, or use mine sorakh2756@gmail.com
+GMAIL_APP_PASSWORD=azktsjaohgvppdxo							# Replace with your actual App Password, or use mine azktsjaohgvppdxo
 EMAIL_FROM_NAME=TopicTrends									# Replace with your desired sender name
-GOOGLE_API_KEY=google-studio-ai-ApiKey						# Replace with your google gemini key, or use mine AIzaSyBfRFz3pcIveQLWea_Sd_JmipPEBieNft4
+GOOGLE_API_KEY=AIzaSyBfRFz3pcIveQLWea_Sd_JmipPEBieNft4# Replace with your google gemini key, or use mine AIzaSyBfRFz3pcIveQLWea_Sd_JmipPEBieNft4
